Extract Every Code worker CLI
